Function openssl_probe::has_ssl_cert_env_vars

pub fn has_ssl_cert_env_vars() -> bool

Check whether the OpenSSL SSL_CERT_FILE and/or SSL_CERT_DIR environment variable is configured in this process with an existing file or directory.

That being the case would indicate that certificates will be found successfully by OpenSSL.

Returns true if either variable is set to an existing file or directory.