Function openssl_probe::try_init_ssl_cert_env_vars

pub fn try_init_ssl_cert_env_vars() -> bool

Probe for SSL certificates on the system, then configure the SSL certificate SSL_CERT_FILE and SSL_CERT_DIR environment variables in this process for OpenSSL to use.

Preconfigured values in the environment variables will not be overwritten if the paths they point to exist and are accessible.

Returns true if any certificate file or directory was found while probing. Combine this with has_ssl_cert_env_vars() to check whether previously configured environment variables are valid.