Crate openssl_probe

Structs

ProbeResult

Constants

ENV_CERT_DIR

The OpenSSL environment variable to configure what certificates directory to use.

ENV_CERT_FILE

The OpenSSL environment variable to configure what certificate file to use.

Functions

find_certs_dirs

Probe the system for the directory in which CA certificates should likely be found.

has_ssl_cert_env_vars

Check whether the OpenSSL SSL_CERT_FILE and/or SSL_CERT_DIR environment variable is configured in this process with an existing file or directory.

init_ssl_cert_env_vars

Probe for SSL certificates on the system, then configure the SSL certificate SSL_CERT_FILE and SSL_CERT_DIR environment variables in this process for OpenSSL to use.

probe

try_init_ssl_cert_env_vars

Probe for SSL certificates on the system, then configure the SSL certificate SSL_CERT_FILE and SSL_CERT_DIR environment variables in this process for OpenSSL to use.